services/api/requirements.txt
fastapi[standard]==0.117.1
pydantic-settings==2.11.0
python-dotenv==1.1.1
sqlmodel==0.0.25
asyncpg==0.30.0
pyjwt==2.10.1
argon2-cffi==23.1.0
redis==6.4.0
aio-pika==9.5.7
cryptography==46.0.3
dspy==3.0.4
langchain-core==1.2.0
langchain-google-genai==4.0.0
langgraph-checkpoint-redis==0.3.0
langchain==1.1.3
# Test dependencies
pytest==8.3.4
pytest-asyncio==1.2.0
httpx==0.28.1
freezegun==1.5.5
